Added `print` as an alias for `get`
The idea is that `get`, `list` and `print` all essentially mean the same thing
and can be used interchangeably. When a user specifies just `get`, `list` or
`print` with no arguments then all entries will be shown omitting passwords.
When a user specifies a service and a username with any of these commands just
that single entry will be shown along with it's password.